[RUMF-1209] introduce rage click chains
If we need to collect rage click, we need to buffer similar clicks
to make sure they don't need to be represented as a single "rage click"
action.

This commit introduce a concept of "click chain", grouping similar
clicks together, and flushing them when we are sure that the click chain
is "complete".