🏷️ [RUMF-1209] add "ES2015" to the test app tsconfig
We now use Set to track actions, and it appears in module type
definitions (.d.ts). By default, TS does not include the type definition
for Set, so we need to use the `ES2015` lib.

This should not be a breaking change, since we already used Set in other
places of the SDK.